Issue #1072: Use 'real(8)' for variable 'rseries' in routine 'ose_calcOmPstddevCH' of module 'obsSpaceErrorStdDev_mod'
This avoid a conversion warning.

/home/erv000/data_maestro/midas/midas-1072-gfortran/src/modules/obsSpaceErrorStdDev_mod.f90:2450:50:

 2450 |             rseries(1:nSeries(levIndex,latIndex))=series(levIndex,latIndex,1:nSeries(levIndex,latIndex))
      |                                                  1
Warning: Possible change of value in conversion from REAL(8) to REAL(4) at (1) [-Wconversion]

+++ src/modules/obsSpaceErrorStdDev_mod.f90
@@ -2305,7 +2305,7 @@ module obsSpaceErrorStdDev_mod
     integer, parameter :: minCount = 5
     real(8), parameter :: stdScale = 2.0
     real(8), parameter :: minVal = 1.0d-20
-    real(4) :: rseries(maxCount)
+    real(8) :: rseries(maxCount)
     integer :: ip(maxCount)
 
     ! Loop over all obs types
